shared/python/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
PyJWT==2.8.0
cryptography==41.0.7
